api/pyproject.toml
[project]
name = "nhl-screener-api"
version = "1.0.0"
requires-python = ">=3.9"
dependencies = [
    "httpx>=0.26.0",
    "beautifulsoup4>=4.12.0",
    "upstash-redis>=1.0.0",
    "pyjwt>=2.8.0",
    "bcrypt>=4.1.0",
]
